refactor: CommonFilePathManager 경로 변경
